Refactored divider to omit default browser styling.
