--- src/components/popup/controls/divider.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components';
-
-export const Divider = styled.hr``;
